Test route predict functionality
