+++ Config/segment_boundaries.txt
@@ -1,0 +1,262 @@
+{
+    "59.0": {
+        "start_coordinates": [
+            39.78110885620117,
+            14.963831901550291
+        ],
+        "end_coordinates": [
+            44.60873031616211,
+            14.470298767089844
+        ]
+    },
+    "56.0": {
+        "start_coordinates": [
+            44.59817123413086,
+            14.463890075683594
+        ],
+        "end_coordinates": [
+            32.35111999511719,
+            16.843290328979492
+        ]
+    },
+    "20.0": {
+        "start_coordinates": [
+            32.2050895690918,
+            17.02786636352539
+        ],
+        "end_coordinates": [
+            31.884401321411133,
+            18.41863441467285
+        ]
+    },
+    "48.0": {
+        "start_coordinates": [
+            31.89437484741211,
+            19.062782287597656
+        ],
+        "end_coordinates": [
+            31.9400634765625,
+            19.90982818603516
+        ]
+    },
+    "52.0": {
+        "start_coordinates": [
+            32.03223419189453,
+            20.53268814086914
+        ],
+        "end_coordinates": [
+            32.24775695800781,
+            21.5244255065918
+        ]
+    },
+    "16.0": {
+        "start_coordinates": [
+            32.26324462890625,
+            21.70364570617676
+        ],
+        "end_coordinates": [
+            32.29417037963867,
+            22.722888946533203
+        ]
+    },
+    "44.0": {
+        "start_coordinates": [
+            32.31187057495117,
+            23.13712310791016
+        ],
+        "end_coordinates": [
+            33.297752380371094,
+            24.75884437561035
+        ]
+    },
+    "12.0": {
+        "start_coordinates": [
+            33.29154586791992,
+            24.74676322937012
+        ],
+        "end_coordinates": [
+            34.99372482299805,
+            25.08304214477539
+        ]
+    },
+    "36.0": {
+        "start_coordinates": [
+            35.40017318725586,
+            25.08298110961914
+        ],
+        "end_coordinates": [
+            38.2685546875,
+            25.076744079589844
+        ]
+    },
+    "4.0": {
+        "start_coordinates": [
+            38.532501220703125,
+            25.070295333862305
+        ],
+        "end_coordinates": [
+            52.71760559082031,
+            26.713008880615234
+        ]
+    },
+    "7.0": {
+        "start_coordinates": [
+            52.725852966308594,
+            26.7133731842041
+        ],
+        "end_coordinates": [
+            38.35975646972656,
+            25.07788848876953
+        ]
+    },
+    "39.0": {
+        "start_coordinates": [
+            38.02962112426758,
+            25.07642936706543
+        ],
+        "end_coordinates": [
+            35.006717681884766,
+            25.08160781860352
+        ]
+    },
+    "15.0": {
+        "start_coordinates": [
+            34.829463958740234,
+            25.077991485595703
+        ],
+        "end_coordinates": [
+            33.34508514404297,
+            24.77989387512207
+        ]
+    },
+    "47.0": {
+        "start_coordinates": [
+            33.19739532470703,
+            24.711162567138672
+        ],
+        "end_coordinates": [
+            32.30157852172852,
+            23.00662040710449
+        ]
+    },
+    "19.0": {
+        "start_coordinates": [
+            32.292781829833984,
+            22.861421585083008
+        ],
+        "end_coordinates": [
+            32.28266143798828,
+            21.9201717376709
+        ]
+    },
+    "55.0": {
+        "start_coordinates": [
+            32.22095489501953,
+            21.35133934020996
+        ],
+        "end_coordinates": [
+            31.95682144165039,
+            20.02157592773437
+        ]
+    },
+    "51.0": {
+        "start_coordinates": [
+            31.93095016479492,
+            19.901945114135746
+        ],
+        "end_coordinates": [
+            31.88814926147461,
+            18.576948165893555
+        ]
+    },
+    "27.0": {
+        "start_coordinates": [
+            31.875459671020508,
+            18.428361892700195
+        ],
+        "end_coordinates": [
+            30.90413475036621,
+            17.35820770263672
+        ]
+    },
+    "63.0": {
+        "start_coordinates": [
+            30.69799041748047,
+            17.351076126098633
+        ],
+        "end_coordinates": [
+            29.374950408935547,
+            18.68052673339844
+        ]
+    },
+    "60.0": {
+        "start_coordinates": [
+            29.37993812561035,
+            18.68202209472656
+        ],
+        "end_coordinates": [
+            30.82598304748535,
+            17.358306884765625
+        ]
+    },
+    "24.0": {
+        "start_coordinates": [
+            31.160064697265625,
+            17.41614532470703
+        ],
+        "end_coordinates": [
+            31.886632919311523,
+            18.467187881469727
+        ]
+    },
+    "10.0": {
+        "start_coordinates": [
+            32.30531311035156,
+            23.08527183532715
+        ],
+        "end_coordinates": [
+            33.57807159423828,
+            23.2078914642334
+        ]
+    },
+    "42.0": {
+        "start_coordinates": [
+            31.88408851623535,
+            25.074922561645508
+        ],
+        "end_coordinates": [
+            29.53225517272949,
+            23.927881240844727
+        ]
+    },
+    "41.0": {
+        "start_coordinates": [
+            29.538101196289062,
+            23.926877975463867
+        ],
+        "end_coordinates": [
+            31.878568649291992,
+            25.087282180786133
+        ]
+    },
+    "9.0": {
+        "start_coordinates": [
+            31.90736198425293,
+            25.035755157470703
+        ],
+        "end_coordinates": [
+            32.30453872680664,
+            23.058448791503903
+        ]
+    },
+    "23.0": {
+        "start_coordinates": [
+            31.887168884277344,
+            18.149568557739254
+        ],
+        "end_coordinates": [
+            32.28999328613281,
+            16.903648376464844
+        ]
+    }
+}
